Code to graph DCI raw data vs ranking
Here is code to read Austin's data (v7) and graph the raw values against the ranking for each of the seven DCI components.  It highlights the San Antonio zip codes.  The shape of the points shows the steep drop from richest zip codes to the large "middle" and then another steep drop to the poorest.  The length of the two steep slopes, and the positions occupied by San Antonio, vary among the DCI components.